unify test mocks
